+++ src/dtos/user.dtos.ts
@@ -1,4 +1,4 @@
-// DTOs and mappers for user, store, review, mission
+// DTOs and mappers for user
 // Converted to TypeScript with explicit interfaces and return types.
 
 // DB User row shape (snake_case) used in responses
@@ -103,128 +103,5 @@ export const responseFromUser = async ({
         point: user.point,
         status: user.status,
         preferences: preferences
-    };
-};
-
-export interface StoreCreateDTO {
-    name: string;
-    food_category_id: number;
-    subscription: string;
-    address: string;
-    detail_address: string;
-}
-
-export const bodyToStore = (body: unknown): StoreCreateDTO => {
-    if (body === null || body === undefined || typeof body !== "object") {
-        throw new Error("body is null or undefined or not an object");
-    }
-    const b = body as any;
-    if (
-        b.name === undefined ||
-        b.food_category_id === undefined ||
-        b.address === undefined ||
-        b.detail_address === undefined
-    ) {
-        throw new Error(
-            "name, food_category_id, address, detail_address is undefined",
-        );
-    } else {
-        return {
-            name: String(b.name),
-            food_category_id: Number(b.food_category_id),
-            subscription: String(b.subscription),
-            address: String(b.address),
-            detail_address: String(b.detail_address),
-        };
-    }
-};
-
-export interface ReviewCreateDTO {
-    member_id: number;
-    store_id: number;
-    grade: string;
-    description: string;
-}
-
-export const bodyToReview = (body: unknown): ReviewCreateDTO => {
-    if (body === null || body === undefined || typeof body !== "object") {
-        throw new Error("body is null or undefined or not an object");
-    }
-    const b = body as any;
-    if (
-        b.member_id === undefined ||
-        b.store_id === undefined ||
-        b.grade === undefined ||
-        b.description === undefined
-    ) {
-        throw new Error(
-            "The following fields are missing: member_id, store_id, grade, description",
-        );
-    }
-    return {
-        member_id: Number(b.member_id),
-        store_id: Number(b.store_id),
-        grade: String(b.grade),
-        description: String(b.description),
-    };
-};
-
-export interface MissionCreateDTO {
-    description: string;
-    title: string;
-    point_reward: number;
-    store_id: number;
-}
-
-export const bodyToMission = (body: unknown): MissionCreateDTO => {
-    if (body === null || body === undefined || typeof body !== "object") {
-        throw new Error("body is null or undefined or not an object");
-    }
-    const b = body as any;
-    if (
-        b.title === undefined ||
-        b.point_reward === undefined ||
-        b.store_id === undefined
-    ) {
-        throw new Error(
-            "The following fields are missing: title, point_reward, store_id",
-        );
-    }
-    return {
-        description: String(b.description),
-        title: String(b.title),
-        point_reward: Number(b.point_reward),
-        store_id: Number(b.store_id),
-    };
-};
-
-export interface MemberMissionCreateDTO {
-    member_id: number;
-    mission_id: number;
-    address: string | null;
-    is_completed: boolean;
-    deadline: string | Date;
-    activated: boolean;
-    created_at: string | Date;
-    updated_at: string | Date;
-}
-
-export const bodyToMemberMission = (body: unknown): MemberMissionCreateDTO => {
-    if (body === null || body === undefined || typeof body !== "object") {
-        throw new Error("body is null or undefined or not an object");
-    }
-    const b = body as any;
-    if (b.member_id === undefined || b.mission_id === undefined) {
-        throw new Error("The following fields are missing: member_id, mission_id");
-    }
-    return {
-        member_id: Number(b.member_id),
-        mission_id: Number(b.mission_id),
-        address: String(b.address) ?? null,
-        is_completed: b.is_completed ?? 0,
-        deadline: b.deadline,
-        activated: b.activated,
-        created_at: b.created_at,
-        updated_at: b.updated_at,
     };
 };
