+++ src/index.ts
@@ -3,19 +3,17 @@ import express, {NextFunction, Request, Response} from "express";
 import cors from "cors";
 import morgan from "morgan";
 import cookieParser from "cookie-parser";
-
+import {handleUserSignUp} from "./controller/user.controller";
+import {handleStoreInsert} from "./controller/store.controller";
+import {handleGetMemberReview, handleGetStoreReivew, handleInsertReview} from "./controller/review.controller";
 import {
-    handleGetMemberReview,
     handleGetOnMission,
     handleGetStoreMission,
-    handleGetStoreReivew,
     handleInsertMission,
-    handleInsertReview,
     handleMissionStart,
-    handleSetMissionCompelete,
-    handleStoreInsert,
-    handleUserSignUp,
-} from "./controller/user.controller";
+    handleSetMissionCompelete
+} from "./controller/mission.controller";
+
 
 dotenv.config();
